move test files to a module folder
