+++ lib/modules/tests/run_tests.js
@@ -1,8 +1,8 @@
 const async = require('async');
 const Mocha = require('mocha');
 const path = require('path');
-const {runCmd} = require('../utils/utils');
-const fs = require('../core/fs');
+const {runCmd} = require('../../utils/utils');
+const fs = require('../../core/fs');
 const assert = require('assert');
 const Test = require('./test');
 const EmbarkSpec = require('./reporter');
