+++ lib/modules/tests/solc_test.js
@@ -1,7 +1,7 @@
 const Test = require('./test');
 const async = require('async');
 const fs = require('fs-extra');
-const File = require('./../core/file');
+const File = require('./../../core/file');
 const remixTests = require('remix-tests');
 const Base = require('mocha/lib/reporters/base');
 const color = Base.color;
