+++ lib/modules/tests/test.js
@@ -1,12 +1,12 @@
 const async = require('async');
-const Engine = require('../core/engine.js');
+const Engine = require('../../core/engine.js');
 const TestLogger = require('./test_logger.js');
 const Web3 = require('web3');
-const Events = require('../core/events');
-const AccountParser = require('../utils/accountParser');
+const Events = require('../../core/events');
+const AccountParser = require('../../utils/accountParser');
 // TODO: breaks module isolation; tests need to be refactored to use the engine and avoid this
-const Provider = require('../modules/blockchain_connector/provider.js');
-const utils = require('../utils/utils');
+const Provider = require('../blockchain_connector/provider.js');
+const utils = require('../../utils/utils');
 
 const EmbarkJS = require('embarkjs');
