+++ lib/modules/tests/test_logger.js
